+++ src/Poker.Service/PlayerCreateArgs.cs
@@ -4,7 +4,7 @@ public record PlayerCreateArgs
 {
     public required Guid Id { get; init; }
 
-    public required double BeginningStack { get; init; }
+    public required uint BeginningStack { get; init; }
 
     public required bool Automaton { get; set; }
 }
